Add deletion policy for projects, update deletion policy naming
