+++ api/v1alpha2/project_types.go
@@ -128,13 +128,11 @@ type ProjectTeamAccess struct {
 // You must use one of the following values:
 // - `retain`: When the custom resource is deleted, the operator will not delete the associated project.
 // - `destroy`: Performs a destroy operation to remove the project. The project must be empty.
-// - `force`: Forcefully and immediately removes all resources in the project. Once this is completed the operator deletes the project.
 type ProjectDeletionPolicy string
 
 const (
 	ProjectDeletionPolicyRetain  ProjectDeletionPolicy = "retain"
 	ProjectDeletionPolicyDestroy ProjectDeletionPolicy = "destroy"
-	ProjectDeletionPolicyForce   ProjectDeletionPolicy = "force"
 )
 
 // ProjectSpec defines the desired state of Project.
@@ -169,10 +167,9 @@ type ProjectSpec struct {
 	// You must use one of the following values:
 	// - `retain`:  When the custom resource is deleted, the operator will not delete the associated project.
 	// - `destroy`: Performs a destroy operation to remove the project. The project must be empty.
-	// - `force`: Forcefully and immediately removes all resources in the project. Once this is completed the operator deletes the project.
 	// Default: `retain`.
 	//
-	//+kubebuilder:validation:Enum:=retain;destroy;force
+	//+kubebuilder:validation:Enum:=retain;destroy
 	//+kubebuilder:default=retain
 	//+optional
 	DeletionPolicy ProjectDeletionPolicy `json:"deletionPolicy,omitempty"`
