+++ config/crd/bases/app.terraform.io_projects.yaml
@@ -60,12 +60,10 @@ spec:
                   You must use one of the following values:
                   - `retain`:  When the custom resource is deleted, the operator will not delete the associated project.
                   - `destroy`: Performs a destroy operation to remove the project. The project must be empty.
-                  - `force`: Forcefully and immediately removes all resources in the project. Once this is completed the operator deletes the project.
                   Default: `retain`.
                 enum:
                 - retain
                 - destroy
-                - force
                 type: string
               name:
                 description: Name of the Project.
