+++ docs/api-reference.md
@@ -448,7 +448,6 @@ DeletionPolicy defines the strategy the Kubernetes operator uses when you delete
 You must use one of the following values:
 - `retain`: When the custom resource is deleted, the operator will not delete the associated project.
 - `destroy`: Performs a destroy operation to remove the project. The project must be empty.
-- `force`: Forcefully and immediately removes all resources in the project. Once this is completed the operator deletes the project.
 
 _Appears in:_
 - [ProjectSpec](#projectspec)
@@ -472,7 +471,7 @@ _Appears in:_
 | `token` _[Token](#token)_ | API Token to be used for API calls. |
 | `name` _string_ | Name of the Project. |
 | `teamAccess` _[ProjectTeamAccess](#projectteamaccess) array_ | HCP Terraform's access model is team-based. In order to perform an action within a HCP Terraform organization,<br />users must belong to a team that has been granted the appropriate permissions.<br />You can assign project-specific permissions to teams.<br />More information:<br />  - https://developer.hashicorp.com/terraform/cloud-docs/workspaces/organize-workspaces-with-projects#permissions<br />  - https://developer.hashicorp.com/terraform/cloud-docs/users-teams-organizations/permissions#project-permissions |
-| `deletionPolicy` _[ProjectDeletionPolicy](#projectdeletionpolicy)_ | DeletionPolicyForProject defines the strategy the Kubernetes operator uses when you delete a project, either manually or by a system event.<br /><br />You must use one of the following values:<br />- `retain`:  When the custom resource is deleted, the operator will not delete the associated project.<br />- `destroy`: Performs a destroy operation to remove the project. The project must be empty.<br />- `force`: Forcefully and immediately removes all resources in the project. Once this is completed the operator deletes the project.<br />Default: `retain`. |
+| `deletionPolicy` _[ProjectDeletionPolicy](#projectdeletionpolicy)_ | DeletionPolicyForProject defines the strategy the Kubernetes operator uses when you delete a project, either manually or by a system event.<br /><br />You must use one of the following values:<br />- `retain`:  When the custom resource is deleted, the operator will not delete the associated project.<br />- `destroy`: Performs a destroy operation to remove the project. The project must be empty.<br />Default: `retain`. |
